Merged PR 158859: Don't use virtual env for backward compatibility with AppService prior implementation
It turns out that the existing app service implementation doesn't activate the virtual env, but rather add the site packages folder from it to the PYTHONPATH variable.
Activating the virtual env doesn't work because of the difference in paths between build and runtime images. So we'll follow that they're doing for apps already built.

Related work items: #762560