Merged PR 164409: Update curl in all dotnet core images that have it installed
Moving 2.2 out of the template, as it is the only one based on Alpine, which doesn't include curl in the first place.
Curl has two security vulnerabilities that were fixed in the latest version. Since most of those versions reached end of life and are no longer supported, the base images themselves won't be updated and we have to fix the issue ourselves.

Related work items: #787236, #792347