Merged PR 168020: Remove packages folder when building with virtual env
Kudu will call Oryx passing the -p virtualenv_name property again so we will go back to virtual environments. This change is to remove packages folder when building with virtual env, so when running the app we will use the virtual env instead of the packages folder.

Related work items: #807896